Resplot tests
